+++ autoarray/dataset/plot/imaging_plotters.py
@@ -5,7 +5,7 @@ from autoarray.plot.include.two_d import Include2D
 from autoarray.plot.mat_plot.two_d import MatPlot2D
 from autoarray.plot.auto_labels import AutoLabels
 from autoarray.plot.abstract_plotters import Plotter
-from autoarray.dataset.imaging.imaging import Imaging
+from autoarray.dataset.imaging.dataset import Imaging
 
 
 class ImagingPlotterMeta(Plotter):
