+++ autoarray/fit/fit_interferometer.py
@@ -1,7 +1,7 @@
 import numpy as np
 from typing import Dict, Optional, Union
 
-from autoarray.dataset.interferometer.interferometer import Interferometer
+from autoarray.dataset.interferometer.dataset import Interferometer
 from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.visibilities import Visibilities
 from autoarray.fit.fit_dataset import FitDataset
