+++ autoarray/inversion/inversion/factory.py
@@ -1,6 +1,6 @@
 from typing import Dict, List, Optional, Union
 
-from autoarray.dataset.imaging.imaging import Imaging
+from autoarray.dataset.imaging.dataset import Imaging
 from autoarray.dataset.imaging.w_tilde import WTildeImaging
 from autoarray.dataset.interferometer.w_tilde import WTildeInterferometer
 from autoarray.structures.arrays.uniform_2d import Array2D
